Add a domUpdates file to manage all display began moving functions over
