+++ src/domUpdates.js
@@ -1,0 +1,51 @@
+const domUpdates = {
+  createUserDisplay(user) {
+  let firstName = user.name.split(" ")[0];
+    let welcomeMsg = `
+      <div class="welcome-msg">
+        <h1>Welcome ${firstName}!</h1>
+      </div>`;
+    document.querySelector(".banner-image").insertAdjacentHTML("afterbegin",
+      welcomeMsg);
+  },
+
+  addToDom(recipeInfo, shortRecipeName, element) {
+  let cardHtml = `
+    <div class="recipe-card" id=${recipeInfo.id}>
+      <h3 maxlength="40">${shortRecipeName}</h3>
+      <div class="card-photo-container">
+        <img src=${recipeInfo.image} class="card-photo-preview" alt="${recipeInfo.name} recipe" title="${recipeInfo.name} recipe">
+        <div class="text">
+          <div>Click for Instructions</div>
+        </div>
+      </div>
+      <h4>${recipeInfo.tags[0]}</h4>
+      <img src="../images/apple-logo-outline.png" alt="unfilled apple icon" class="card-apple-icon">
+    </div>`
+   element.insertAdjacentHTML("beforeend", cardHtml);
+  },
+  
+  listTags(allTags, element) {
+    allTags.forEach(tag => {
+      let tagHtml = `<li><input type="checkbox" class="checked-tag" id="${tag}">
+        <label for="${tag}">${this.capitalize(tag)}</label></li>`;
+      element.insertAdjacentHTML("beforeend", tagHtml);
+    });
+  },
+  
+  capitalize(words) {
+    return words.split(" ").map(word => {
+      return word.charAt(0).toUpperCase() + word.slice(1);
+    }).join(" ");
+  },
+
+  showAllRecipes(allRecipes) {
+    allRecipes.forEach(recipe => {
+      let domRecipe = document.getElementById(`${recipe.id}`);
+      domRecipe.style.display = "block";
+    });
+    // showWelcomeBanner();
+  }
+};
+
+export default domUpdates
