+++ src/scripts.js
@@ -4,6 +4,7 @@ import './css/styles.scss';
 import User from './user';
 import Recipe from './recipe';
 import ApiFetch from './ApiFetch'
+import domUpdates from './domUpdates'
 
 let api = new ApiFetch();
 
@@ -44,7 +45,7 @@ let tagList = document.querySelector(".tag-list");
 
 
 // ON CLICK EVENTS
-allRecipesBtn.addEventListener("click", showAllRecipes);
+allRecipesBtn.addEventListener("click", domUpdates.showAllRecipes(recipes));
 filterBtn.addEventListener("click", findCheckedBoxes);
 main.addEventListener("click", addToMyRecipes);
 pantryBtn.addEventListener("click", toggleMenu);
@@ -55,17 +56,7 @@ showPantryRecipes.addEventListener("click", findCheckedPantryBoxes);
 function generateUser(users, ingredients) {
   user = new User(users[Math.floor(Math.random() * users.length)]);
   findPantryInfo(ingredients);
-  createUserDisplay(user);
-}
-
-function createUserDisplay(user) {
-let firstName = user.name.split(" ")[0];
-  let welcomeMsg = `
-    <div class="welcome-msg">
-      <h1>Welcome ${firstName}!</h1>
-    </div>`;
-  document.querySelector(".banner-image").insertAdjacentHTML("afterbegin",
-    welcomeMsg);
+  domUpdates.createUserDisplay(user);
 }
 
 // CREATE RECIPE CARDS
@@ -77,24 +68,8 @@ function createCards(recipeData) {
     if (recipeInfo.name.length > 40) {
       shortRecipeName = recipeInfo.name.substring(0, 40) + "...";
     }
-    addToDom(recipeInfo, shortRecipeName)
+    domUpdates.addToDom(recipeInfo, shortRecipeName, main)
   });
-}
-
-function addToDom(recipeInfo, shortRecipeName) {
-  let cardHtml = `
-    <div class="recipe-card" id=${recipeInfo.id}>
-      <h3 maxlength="40">${shortRecipeName}</h3>
-      <div class="card-photo-container">
-        <img src=${recipeInfo.image} class="card-photo-preview" alt="${recipeInfo.name} recipe" title="${recipeInfo.name} recipe">
-        <div class="text">
-          <div>Click for Instructions</div>
-        </div>
-      </div>
-      <h4>${recipeInfo.tags[0]}</h4>
-      <img src="../images/apple-logo-outline.png" alt="unfilled apple icon" class="card-apple-icon">
-    </div>`
-  main.insertAdjacentHTML("beforeend", cardHtml);
 }
 
 // FILTER BY RECIPE TAGS
@@ -108,22 +83,10 @@ function findTags(recipeData) {
     });
   });
   tags.sort();
-  listTags(tags);
+  domUpdates.listTags(tags, tagList);
 }
 
-function listTags(allTags) {
-  allTags.forEach(tag => {
-    let tagHtml = `<li><input type="checkbox" class="checked-tag" id="${tag}">
-      <label for="${tag}">${capitalize(tag)}</label></li>`;
-    tagList.insertAdjacentHTML("beforeend", tagHtml);
-  });
-}
 
-function capitalize(words) {
-  return words.split(" ").map(word => {
-    return word.charAt(0).toUpperCase() + word.slice(1);
-  }).join(" ");
-}
 
 function findCheckedBoxes() {
   let tagCheckboxes = document.querySelectorAll(".checked-tag");
@@ -146,7 +109,7 @@ function findTaggedRecipes(selected) {
       }
     })
   });
-  showAllRecipes();
+  domUpdates.showAllRecipes(recipes);
   if (filteredResults.length > 0) {
     filterRecipes(filteredResults);
   }
@@ -276,13 +239,6 @@ function toggleMenu() {
   }
 }
 
-function showAllRecipes() {
-  recipes.forEach(recipe => {
-    let domRecipe = document.getElementById(`${recipe.id}`);
-    domRecipe.style.display = "block";
-  });
-  showWelcomeBanner();
-}
 
 
 // CREATE AND USE PANTRY
@@ -320,7 +276,7 @@ function findCheckedPantryBoxes() {
   let selectedIngredients = pantryCheckboxInfo.filter(box => {
     return box.checked;
   })
-  showAllRecipes();
+  domUpdates.showAllRecipes(recipes);
   if (selectedIngredients.length > 0) {
     findRecipesWithCheckedIngredients(selectedIngredients);
   }
